Added label2skeleton mapping. Added skeleton sliding window filtering. Added new local bfs algorithm. Added tests for all new methods.

+++ morphx/processing/hybrids.py
@@ -62,7 +62,7 @@ def extract_mesh_subset(hm: HybridMesh, local_bfs: np.ndarray) -> MeshCloud:
         new_labels = hm.labels
     else:
         new_faces = hm.faces[total_face]
-        total_vertex = np.unique(new_faces.flatten())
+        total_vertex = np.unique(new_faces.flatten()).astype(int)
         new_vertices = hm.vertices[total_vertex]
         new_labels = hm.labels[total_vertex]
         # normalize new faces to be contiguous
